ISIS-3051: fighting lombok issues
